Bugs fixing (#34)
* UI bugs fixes

* Second car choose fix & size fix

* make bottom left UI indicator

* Updated readme

* make all UI

* ui color fix

* fix 2 player UI

* fix network, fix map

* fix cpplint

* Update screenshots

* fix bots, fix shooting range

* Bugs & style fixes

* Minifix

* Credits

* Credits fix

* Keyboards manual

* Superfix

* Background & map widget addons

* Changing some hp and bonuses constants

* Window size fix

* Freeze fix

* Freeze fix №2

* Network tiles fix

* Update README.md

+++ src/helpers/cars_data.h
@@ -11,7 +11,7 @@ struct CarsData {
   explicit CarsData(std::vector<CarAchievements> new_cars_data);
 
   size_t GetVelocity(int index) const;
-  size_t GetLapsCounter(int index) const;
+  int32_t GetLapsCounter(int index) const;
   size_t GetBulletsAmount(int index) const;
   size_t GetMinesAmount(int index) const;
   size_t GetNumber(int index) const;
