+++ src/helpers/game_mode.h
@@ -7,7 +7,7 @@ struct GameMode {
   size_t players_amount = 1;
   size_t network_players_amount = 0;
   size_t bots_amount = 0;
-  size_t laps_amount = 1;
+  int32_t laps_amount = 1;
   size_t first_player_car_number = 0;
   size_t second_player_car_number = 0;
   int volume_settings_parameter = 0;
